Remove unused Bug import
