+++ src/components/ui/badge.tsx
@@ -1,11 +1,7 @@
 import * as React from 'react'
-import { type VariantProps } from 'class-variance-authority'
 import { cn } from '../../lib/utils'
 import { badgeVariants } from './badge-variants'
-
-export interface BadgeProps
-  extends React.HTMLAttributes<HTMLDivElement>,
-    VariantProps<typeof badgeVariants> {}
+import type { BadgeProps } from '../../types/ui/badge'
 
 function Badge({ className, variant, ...props }: BadgeProps) {
   return (
