+++ src/routes/Leaderboard.tsx
@@ -3,6 +3,7 @@ import { Link } from 'react-router-dom'
 import { useBugStore } from '../store'
 import Meta from '../components/Meta'
 import { Input } from '../components/ui/input'
+import type { SortKey } from '../types/routes/Leaderboard'
 
 /** Map bounty → tier name */
 const levelFromBounty = (bounty: number): string => {
@@ -19,8 +20,6 @@ const levelTier = (bounty: number): number => {
   if (bounty >= 2000) return 2
   return 1
 }
-
-type SortKey = 'rank' | 'name' | 'bugs' | 'bounty' | 'efficiency' | 'level'
 
 export default function Leaderboard() {
   const users = useBugStore(s => s.users)
