+++ src/types/components/AimCursor.ts
@@ -1,0 +1,4 @@
+export interface AimCursorProps {
+  x: number
+  y: number
+}
