+++ src/types/components/BugTrendsChart.ts
@@ -1,0 +1,5 @@
+export interface DataPoint {
+  date: Date
+  created: number
+  resolved: number
+}
