+++ src/types/components/Taskbar.ts
@@ -1,0 +1,5 @@
+export interface TaskbarProps {
+  windowTitle: string
+  minimized: boolean
+  onToggle: () => void
+}
