+++ src/types/routes/Leaderboard.ts
@@ -1,0 +1,7 @@
+export type SortKey =
+  | 'rank'
+  | 'name'
+  | 'bugs'
+  | 'bounty'
+  | 'efficiency'
+  | 'level'
