+++ src/types/ui/input.ts
@@ -1,0 +1,4 @@
+import * as React from 'react'
+
+export interface InputProps
+  extends React.InputHTMLAttributes<HTMLInputElement> {}
