+++ src/types/ui/label.ts
@@ -1,0 +1,4 @@
+import * as React from 'react'
+
+export interface LabelProps
+  extends React.LabelHTMLAttributes<HTMLLabelElement> {}
