+++ src/types/ui/textarea.ts
@@ -1,0 +1,4 @@
+import * as React from 'react'
+
+export interface TextareaProps
+  extends React.TextareaHTMLAttributes<HTMLTextAreaElement> {}
